refactor(test): split 9 oversized test files into partial classes
CLAUDE.md caps any *.cs at 300 lines. The 9 files below all exceeded
that cap; each is now split into 2-4 partial-class files grouped by
behaviour. xUnit discovers tests across partials, helpers stay in
exactly one file, and all 9 source files re-test green individually.

- ShoppingLedgerTests (598 → main + ShoppingMode + UndoFreezeAtHome)
- WeeklyPlanTests (597 → main + ExtendedMode + ContentTypes + Lifecycle)
- RecipeTests (532 → main + Update + DeleteAndTags)
- AspireFixture (444 → main + Db + Auth)
- GetRecipesQueryHandlerTests (428 → main + Mapping)
- ShoppingEventStoreTests (411 → main + LoadAndConstraints + ModuleEvents)
- SemanticKernelRecipeExtractionServiceTests (420 → main + PromptAndCache)
- SaveRecipeValidatorTests (456 → main + OptionalFields + StringAndEdgeCases)
- SemanticKernelChatServiceFunctionCallingTests (346 → main + Helpers)

Every partial class declaration is wrapped in
#pragma warning disable/restore SA1601 — same pattern the existing
Architecture.Tests partials use — since StyleCop requires XML docs on
every partial declaration and the test-class names are self-explanatory.

No test logic changed. Partial-class names mirror the existing
KeycloakAdminService.Delete.cs / .GetEmail.cs convention so the file
list stays readable. Each new file's using directives are scoped to
what its tests actually need rather than re-inheriting the main's
full block.
