+++ tests/Yumney.Shopping.Domain.Tests/ShoppingLedger/ShoppingLedgerTests.cs
@@ -9,7 +9,9 @@ using Xunit;
 
 namespace SmartSolutionsLab.Yumney.Shopping.Domain.Tests.ShoppingLedger;
 
-public class ShoppingLedgerTests
+#pragma warning disable SA1601
+public partial class ShoppingLedgerTests
+#pragma warning restore SA1601
 {
 	[Fact]
 	public void Create_ValidOwner_ReturnsEmptyLedger()
@@ -251,342 +253,6 @@ public class ShoppingLedgerTests
 		var act = () => ledger.MarkBought(N(itemName!), Q(1, "L"));
 
 		act.Should().Throw<GuardException>();
-	}
-
-	[Fact]
-	public void StartShoppingMode_SetsIsInShoppingMode()
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-
-		ledger.StartShoppingMode();
-
-		ledger.IsInShoppingMode.Should().BeTrue();
-		ledger.ShoppingModeStartedAt.Should().NotBeNull();
-		ledger.PendingChangesCount.Should().Be(0);
-	}
-
-	[Fact]
-	public void StartShoppingMode_AlreadyInMode_NoOp()
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.StartShoppingMode();
-		var version = ledger.Version;
-
-		ledger.StartShoppingMode();
-
-		ledger.Version.Should().Be(version);
-	}
-
-	[Fact]
-	public void EndShoppingMode_ClearsShoppingState()
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.StartShoppingMode();
-
-		ledger.EndShoppingMode(acceptPendingChanges: true);
-
-		ledger.IsInShoppingMode.Should().BeFalse();
-		ledger.ShoppingModeStartedAt.Should().BeNull();
-	}
-
-	[Fact]
-	public void EndShoppingMode_NotInMode_NoOp()
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		var version = ledger.Version;
-
-		ledger.EndShoppingMode(acceptPendingChanges: false);
-
-		ledger.Version.Should().Be(version);
-	}
-
-	[Fact]
-	public void AddItem_WhileInShoppingMode_IncrementsPendingChanges()
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.AddItem(N("Milk"), Q(1, "L"), ItemSource.Manual);
-		ledger.StartShoppingMode();
-
-		ledger.AddItem(N("Eggs"), Q(6, "pc"), ItemSource.Manual);
-		ledger.AddItem(N("Bread"), Q(1, "pc"), ItemSource.Manual);
-
-		ledger.PendingChangesCount.Should().Be(2);
-	}
-
-	[Fact]
-	public void EndShoppingMode_ResetsPendingChanges()
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.StartShoppingMode();
-		ledger.AddItem(N("Eggs"), Q(6, "pc"), ItemSource.Manual);
-
-		ledger.EndShoppingMode(acceptPendingChanges: false);
-
-		ledger.PendingChangesCount.Should().Be(0);
-	}
-
-	[Fact]
-	public void RemoveItem_WhileInShoppingMode_IncrementsPendingChanges()
-	{
-		var milk = N("Milk");
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.AddItem(milk, Q(2, "L"), ItemSource.Manual);
-		ledger.StartShoppingMode();
-
-		ledger.RemoveItem(milk, Q(1, "L"));
-
-		ledger.PendingChangesCount.Should().Be(1);
-	}
-
-	[Fact]
-	public void AdjustQuantity_WhileInShoppingMode_IncrementsPendingChanges()
-	{
-		var milk = N("Milk");
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.AddItem(milk, Q(2, "L"), ItemSource.Manual);
-		ledger.StartShoppingMode();
-
-		ledger.AdjustQuantity(milk, Q(5, "L"));
-
-		ledger.PendingChangesCount.Should().Be(1);
-	}
-
-	[Fact]
-	public void MarkBought_WhileInShoppingMode_DoesNotIncrementPendingChanges()
-	{
-		var milk = N("Milk");
-		var quantity = Q(2, "L");
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.AddItem(milk, quantity, ItemSource.Manual);
-		ledger.StartShoppingMode();
-
-		ledger.MarkBought(milk, quantity);
-
-		ledger.PendingChangesCount.Should().Be(0);
-	}
-
-	[Fact]
-	public void AddItem_NotInShoppingMode_DoesNotTrackPending()
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-
-		ledger.AddItem(N("Milk"), Q(1, "L"), ItemSource.Manual);
-
-		ledger.PendingChangesCount.Should().Be(0);
-	}
-
-	[Fact]
-	public void FromEvents_RebuildShoppingModeState()
-	{
-		var owner = Owner("user-123");
-		var original = Domain.ShoppingLedger.ShoppingLedger.Create(owner);
-		original.AddItem(N("Milk"), Q(1, "L"), ItemSource.Manual);
-		original.StartShoppingMode();
-		original.AddItem(N("Eggs"), Q(6, "pc"), ItemSource.Manual);
-
-		var events = original.UncommittedEvents.ToList();
-		var rebuilt = Domain.ShoppingLedger.ShoppingLedger.FromEvents(original.Identifier, owner, events);
-
-		rebuilt.IsInShoppingMode.Should().BeTrue();
-		rebuilt.ShoppingModeStartedAt.Should().NotBeNull();
-		rebuilt.PendingChangesCount.Should().Be(1);
-	}
-
-	[Fact]
-	public void FromEvents_RebuildEndedShoppingMode()
-	{
-		var owner = Owner("user-123");
-		var original = Domain.ShoppingLedger.ShoppingLedger.Create(owner);
-		original.StartShoppingMode();
-		original.AddItem(N("Eggs"), Q(6, "pc"), ItemSource.Manual);
-		original.EndShoppingMode(acceptPendingChanges: true);
-
-		var events = original.UncommittedEvents.ToList();
-		var rebuilt = Domain.ShoppingLedger.ShoppingLedger.FromEvents(original.Identifier, owner, events);
-
-		rebuilt.IsInShoppingMode.Should().BeFalse();
-		rebuilt.PendingChangesCount.Should().Be(0);
-	}
-
-	[Fact]
-	public void StartShoppingMode_RaisesCorrectEvent()
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-
-		ledger.StartShoppingMode();
-
-		ledger.UncommittedEvents.Should().HaveCount(1);
-		ledger.UncommittedEvents.First().Should().BeOfType<ShoppingModeStarted>();
-	}
-
-	[Fact]
-	public void EndShoppingMode_RaisesCorrectEvent()
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.StartShoppingMode();
-
-		ledger.EndShoppingMode(acceptPendingChanges: false);
-
-		ledger.UncommittedEvents.Should().HaveCount(2);
-		var endEvent = ledger.UncommittedEvents.Last() as ShoppingModeEnded;
-		endEvent.Should().NotBeNull();
-		endEvent!.AcceptedPendingChanges.Should().BeFalse();
-	}
-
-	[Fact]
-	public void UndoBought_ReversesBoughtQuantity()
-	{
-		var milk = N("Milk");
-		var quantity = Q(2, "L");
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.AddItem(milk, quantity, ItemSource.Manual);
-		ledger.MarkBought(milk, quantity);
-
-		ledger.UndoBought(milk, quantity);
-
-		var item = ledger.Items.Values.First();
-		item.Bought.Should().Be(Amount.From(0));
-		item.IsBought.Should().BeFalse();
-	}
-
-	[Fact]
-	public void UndoBought_PartialUndo()
-	{
-		var milk = N("Milk");
-		var quantity = Q(3, "L");
-		var undoQuantity = Q(1, "L");
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.AddItem(milk, quantity, ItemSource.Manual);
-		ledger.MarkBought(milk, quantity);
-
-		ledger.UndoBought(milk, undoQuantity);
-
-		ledger.Items.Values.First().Bought.Should().Be(Amount.From(2));
-	}
-
-	[Fact]
-	public void UndoBought_ExceedsBought_ThrowsBusinessRuleValidation()
-	{
-		var milk = N("Milk");
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.AddItem(milk, Q(1, "L"), ItemSource.Manual);
-
-		var act = () => ledger.UndoBought(milk, Q(5, "L"));
-
-		act.Should().Throw<BusinessRuleValidationException>()
-			.WithMessage("Cannot undo more than the recorded bought quantity.");
-	}
-
-	[Theory]
-	[InlineData(null)]
-	[InlineData("")]
-	public void UndoBought_EmptyItemName_ThrowsGuardException(string? itemName)
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-
-		var act = () => ledger.UndoBought(N(itemName!), Q(1, "L"));
-
-		act.Should().Throw<GuardException>();
-	}
-
-	[Fact]
-	public void AddAsAtHome_AddsDirectlyToBought()
-	{
-		var butter = N("Butter");
-		var quantity = Q(250, "g");
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-
-		ledger.AddAsAtHome(butter, quantity);
-
-		var item = ledger.Items.Values.First();
-		item.Bought.Should().Be(quantity.Amount);
-		item.OnList.Should().Be(Amount.From(0));
-		item.AtHome.Should().Be(quantity.Amount);
-	}
-
-	[Theory]
-	[InlineData(null)]
-	[InlineData("")]
-	public void AddAsAtHome_EmptyItemName_ThrowsGuardException(string? itemName)
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-
-		var act = () => ledger.AddAsAtHome(N(itemName!), Q(1, "pc"));
-
-		act.Should().Throw<GuardException>();
-	}
-
-	[Fact]
-	public void FromEvents_RebuildsUndoBoughtState()
-	{
-		var owner = Owner("user-123");
-		var milk = N("Milk");
-		var quantity = Q(2, "L");
-		var original = Domain.ShoppingLedger.ShoppingLedger.Create(owner);
-		original.AddItem(milk, quantity, ItemSource.Manual);
-		original.MarkBought(milk, quantity);
-		original.UndoBought(milk, quantity);
-
-		var events = original.UncommittedEvents.ToList();
-		var rebuilt = Domain.ShoppingLedger.ShoppingLedger.FromEvents(original.Identifier, owner, events);
-
-		rebuilt.Items.Values.First().Bought.Should().Be(Amount.From(0));
-	}
-
-	[Fact]
-	public void FromEvents_RebuildsAddAsAtHomeState()
-	{
-		var owner = Owner("user-123");
-		var butter = N("Butter");
-		var quantity = Q(250, "g");
-		var original = Domain.ShoppingLedger.ShoppingLedger.Create(owner);
-		original.AddAsAtHome(butter, quantity);
-
-		var events = original.UncommittedEvents.ToList();
-		var rebuilt = Domain.ShoppingLedger.ShoppingLedger.FromEvents(original.Identifier, owner, events);
-
-		rebuilt.Items.Values.First().AtHome.Should().Be(quantity.Amount);
-	}
-
-	[Fact]
-	public void MarkAsFrozen_RaisesEventWithoutTouchingState()
-	{
-		var milk = N("Milk");
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-		ledger.AddAsAtHome(milk, Q(2, "L"));
-		var atHomeBefore = ledger.Items.Values.Single().AtHome;
-
-		ledger.MarkAsFrozen(milk, Unit.From("L"));
-
-		ledger.UncommittedEvents.OfType<ShoppingItemMarkedAsFrozen>().Should().ContainSingle();
-		ledger.Items.Values.Single().AtHome.Should().Be(atHomeBefore);
-	}
-
-	[Fact]
-	public void MarkAsFrozen_NullUnit_RaisesEventWithNullUnit()
-	{
-		var ledger = Domain.ShoppingLedger.ShoppingLedger.Create(Owner("user-123"));
-
-		ledger.MarkAsFrozen(N("Eggs"), null);
-
-		var raised = ledger.UncommittedEvents.OfType<ShoppingItemMarkedAsFrozen>().Single();
-		raised.Unit.Should().BeNull();
-	}
-
-	[Fact]
-	public void FromEvents_RebuildsMarkAsFrozenWithoutCorruptingState()
-	{
-		var owner = Owner("user-123");
-		var butter = N("Butter");
-		var quantity = Q(250, "g");
-		var original = Domain.ShoppingLedger.ShoppingLedger.Create(owner);
-		original.AddAsAtHome(butter, quantity);
-		original.MarkAsFrozen(butter, Unit.From("g"));
-
-		var events = original.UncommittedEvents.ToList();
-		var rebuilt = Domain.ShoppingLedger.ShoppingLedger.FromEvents(original.Identifier, owner, events);
-
-		rebuilt.Items.Values.Single().AtHome.Should().Be(quantity.Amount);
 	}
 
 	private static ItemName N(string name) => ItemName.From(name);
